+++ src/entity/Insumo.java
@@ -4,20 +4,18 @@ public class Insumo {
 	
 	private int IdInsumo;
 	private String nombre;
-	private int precio;
-	public Insumo(int idInsumo, String nombre, int precio) {
+	private double precio;
+	public Insumo(int idInsumo, String nombre, double precio) {
 		super();
 		IdInsumo = idInsumo;
 		this.nombre = nombre;
 		this.precio = precio;
 	}
-
-	public Insumo(String nombre, int precio) {
+	public Insumo(String nombre, double precio) {
 		super();
 		this.nombre = nombre;
 		this.precio = precio;
 	}
-
 	public int getIdInsumo() {
 		return IdInsumo;
 	}
@@ -30,18 +28,16 @@ public class Insumo {
 	public void setNombre(String nombre) {
 		this.nombre = nombre;
 	}
-	public int getPrecio() {
+	public double getPrecio() {
 		return precio;
 	}
-	public void setPrecio(int precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
 	@Override
 	public String toString() {
 		return "Insumo [IdInsumo=" + IdInsumo + ", nombre=" + nombre + ", precio=" + precio + "]";
 	}
-	
-	
-	
+
 	
 }
